Allow OP to see his own hidden content
Original extension works on the assumption that replies is the only
trigger for showing or hiding. So if a person posts (his openning post)
he is automatically shown the hidden content.
There was no check to see if current user is the OP for unhiding.

This fork allows someone to reveal based on only thanks.
So must add a code to check if current user is the OP.